move more pack code in pack.rs
